Created a circle
One circle = one particle                                                                                                          "Context.arc()" needs five coordinates. The first two are the x and y coordinates. The third is the radius of the circle. The last two are the start and end angles.                           "Context.arc()" can be used to create a curved line or a semi-circle.                                       Since it can be used to draw lines, we need to call "context.beginPath()". This lets JavaScript know that we are not starting a new shape that is not connected to any previous lines, if there are any.                                                                                                          "context.strokeStyle" creates an outline of a circle. "context.lineWidth" determines how thick the outline is.                                                                                                          "context.fill()" and "context.stroke()" displays the circle on the page.