Created paintbrush effect
X and Y coordinates for the mouse object are undefined so that the browser loads a blank canvas.                                                                                                                               Added "mouse.x" and "mouse.y" to "context.arc" so that the coordinates are updated when I move my mouse over the window.                                                                                         "mouse.x" is equal to "event.x" so that the browser coordinate can be stored in my custom mouse object for global use.